Writing unit tests for express routes

+++ routes/evidenceFunctions.js
@@ -1,13 +1,13 @@
-const { formatSetClauseValue } = require("../util");
+import { formatSetClauseValue } from "../util.js";
 
-const getEvidenceExpressFunc =
+export const getEvidenceExpressFunc =
   (itemCollectionName, itemIndexKey) => (req, res, next) => {
     const { product, [itemIndexKey]: ix } = req;
 
     return res.json(product[itemCollectionName][ix].evidence);
   };
 
-const addEvidenceExpressFunc = (itemIdKey) => async (req, res, next) => {
+export const addEvidenceExpressFunc = (itemIdKey) => async (req, res, next) => {
   const { [itemIdKey]: id, body: record } = req;
 
   const fields = Object.keys(record);
@@ -27,33 +27,39 @@ const addEvidenceExpressFunc = (itemIdKey) => async (req, res, next) => {
   });
 };
 
-const updateEvidenceExpressFunc = (itemIdKey) => async (req, res, next) => {
-  const { [itemIdKey]: itemId, body: records } = req;
-  let evidence;
-  if (records.length === 0) {
-    await req.client.query({
-      text: "delete from evidence where story_id = $1::integer",
-      values: [itemId],
-    });
-  } else {
-    await Promise.all(
-      records.map(async (record) => {
-        if (record.id) {
-          evidence = record;
-          const setClauseItems = Object.keys(record)
-            .filter(
-              (key) => key !== "id" && key !== "trends" && key !== "story_id"
-            )
-            .map((recordKey) => `${recordKey} = '${record[recordKey]}'`);
-          setClauseItems.push(`${itemIdKey} = ${itemId}`);
-          const setClause = setClauseItems.join(", ");
-          return req.client.query({
-            text: `update evidence set ${setClause} where id = $1::integer`,
-            values: [record.id],
-          });
-        } else {
-          const fields = Object.keys(record).filter((key) => key !== "trends");
-          const query = `insert into evidence 
+export const updateEvidenceExpressFunc =
+  (itemIdKey) => async (req, res, next) => {
+    const { [itemIdKey]: itemId, body: records } = req;
+    if (!itemId || !records) {
+      next("Missing arguments");
+    }
+    let evidence;
+    if (records.length === 0) {
+      await req.client.query({
+        text: "delete from evidence where story_id = $1::integer",
+        values: [itemId],
+      });
+    } else {
+      await Promise.all(
+        records.map(async (record) => {
+          if (record.id) {
+            evidence = record;
+            const setClauseItems = Object.keys(record)
+              .filter(
+                (key) => key !== "id" && key !== "trends" && key !== "story_id"
+              )
+              .map((recordKey) => `${recordKey} = '${record[recordKey]}'`);
+            setClauseItems.push(`${itemIdKey} = ${itemId}`);
+            const setClause = setClauseItems.join(", ");
+            return req.client.query({
+              text: `update evidence set ${setClause} where id = $1::integer`,
+              values: [record.id],
+            });
+          } else {
+            const fields = Object.keys(record).filter(
+              (key) => key !== "trends"
+            );
+            const query = `insert into evidence 
                   (${fields.join(
                     ", "
                   )}, ${itemIdKey}, created_date, modified_date) 
@@ -61,52 +67,46 @@ const updateEvidenceExpressFunc = (itemIdKey) => async (req, res, next) => {
                     formatSetClauseValue(record[field])
                   )}, $1::integer, current_timestamp, current_timestamp)
                   returning *`;
-          console.log("*** query: ", query);
-          evidence = await req.client
-            .query({
-              text: query,
-              values: [itemId],
-            })
-            .then((result) => result.rows[0]);
-        }
-      })
-    );
-    console.log("*** evidence: ", evidence);
-    const trends = evidence.trends ?? [];
-    const existingTrends = await req.client
-      .query({
-        text: "select * from trends where evidence_id = $1::integer",
-        values: [evidence.id],
-      })
-      .then((result) => result.rows);
-    console.log("*** existingTrends: ", JSON.stringify(existingTrends));
-    await Promise.all(
-      trends.map(async (trend) => {
-        const etrend = existingTrends.find((t) => t.name === trend.name);
-        console.log("*** etrend: ", etrend);
-        if (etrend) {
-          await req.client.query({
-            text: `update trends set type = $1::text where id = $2::integer`,
-            values: [trend.type, etrend.id],
-          });
-          console.log("*** updated existing trend: ", etrend.id, trend.type);
-        } else {
-          await req.client.query({
-            text: "insert into trends (name, type, evidence_id) values ($1::text, $2::text, $3::integer)",
-            values: [trend.name, trend.type, evidence.id],
-          });
-          console.log("*** inserted new trend: ", trend.name, trend.type);
-        }
-      })
-    );
-  }
-  req.client.release();
-  return res.json({
-    success: true,
-  });
-};
+            evidence = await req.client
+              .query({
+                text: query,
+                values: [itemId],
+              })
+              .then((result) => result.rows[0]);
+          }
+        })
+      );
+      const trends = evidence.trends ?? [];
+      const existingTrends = await req.client
+        .query({
+          text: "select * from trends where evidence_id = $1::integer",
+          values: [evidence.id],
+        })
+        .then((result) => result.rows);
+      await Promise.all(
+        trends.map(async (trend) => {
+          const etrend = existingTrends.find((t) => t.name === trend.name);
+          if (etrend) {
+            await req.client.query({
+              text: `update trends set type = $1::text where id = $2::integer`,
+              values: [trend.type, etrend.id],
+            });
+          } else {
+            await req.client.query({
+              text: "insert into trends (name, type, evidence_id) values ($1::text, $2::text, $3::integer)",
+              values: [trend.name, trend.type, evidence.id],
+            });
+          }
+        })
+      );
+    }
+    req.client.release();
+    return res.json({
+      success: true,
+    });
+  };
 
-const trackEvidenceIndexExpressFunc =
+export const trackEvidenceIndexExpressFunc =
   (itemCollectionName, itemIndexKey) => (req, res, next) => {
     const { evidence_ix } = req.params;
     const { product, [itemIndexKey]: itemIx } = req;
@@ -122,7 +122,7 @@ const trackEvidenceIndexExpressFunc =
     return next(err);
   };
 
-const deleteEvidenceExpressFunc =
+export const deleteEvidenceExpressFunc =
   (itemCollectionName, itemIndexKey) => async (req, res, next) => {
     const { product, [itemIndexKey]: itemIndex } = req;
     product[itemCollectionName][itemIndex].evidence.splice(req.evidence_ix, 1);
@@ -132,7 +132,7 @@ const deleteEvidenceExpressFunc =
 
 // *** trends functions ***
 
-const getTrendsExpressFunc =
+export const getTrendsExpressFunc =
   (itemCollectionName, itemIndexKey) => (req, res, next) => {
     const { product, [itemIndexKey]: itemIndex } = req;
     return res.json(
@@ -140,7 +140,7 @@ const getTrendsExpressFunc =
     );
   };
 
-const addTrendExpressFunc =
+export const addTrendExpressFunc =
   (itemCollectionName, itemIndexKey) => async (req, res, next) => {
     const { product, [itemIndexKey]: itemIndex } = req;
     var trend = req.body;
@@ -162,7 +162,7 @@ const addTrendExpressFunc =
     return res.json(trend);
   };
 
-const changeTrendExpressFunc =
+export const changeTrendExpressFunc =
   (itemCollectionName, itemIndexKey) => async (req, res, next) => {
     const { product, [itemIndexKey]: itemIndex } = req;
     var evIx = req.evidence_ix;
@@ -179,7 +179,7 @@ const changeTrendExpressFunc =
     });
   };
 
-const deleteTrendExpressFunc =
+export const deleteTrendExpressFunc =
   (itemCollectionName, itemIndexKey) => async (req, res, next) => {
     const { product, [itemIndexKey]: itemIndex } = req;
     var evIx = req.evidence_ix;
@@ -196,14 +196,14 @@ const deleteTrendExpressFunc =
     });
   };
 
-module.exports = {
-  getEvidenceExpressFunc,
-  addEvidenceExpressFunc,
-  trackEvidenceIndexExpressFunc,
-  updateEvidenceExpressFunc,
-  deleteEvidenceExpressFunc,
-  getTrendsExpressFunc,
-  addTrendExpressFunc,
-  changeTrendExpressFunc,
-  deleteTrendExpressFunc,
-};
+// module.exports = {
+//   getEvidenceExpressFunc,
+//   addEvidenceExpressFunc,
+//   trackEvidenceIndexExpressFunc,
+//   updateEvidenceExpressFunc,
+//   deleteEvidenceExpressFunc,
+//   getTrendsExpressFunc,
+//   addTrendExpressFunc,
+//   changeTrendExpressFunc,
+//   deleteTrendExpressFunc,
+// };
